services/order-service/requirements.txt
#TODO: Completare
flask
sqlalchemy==2.0.21
psycopg2-binary
flask-jwt-extended
bcrypt
kafka-python